+++ pkg/api/v1alpha1/jobexecution_types.go
@@ -40,32 +40,28 @@ type JobExecutionSpec struct {
 
 // JobExecutionStatus defines the observed state of JobExecution
 type JobExecutionStatus struct {
-	// Phase has the current state of the Job, it could be one of:
-	// - "Invalid": The JobExecution is referring to a non-existent JobTemplate;
-	// - "Waiting": Waiting to be scheduled;
-	// - "Active": The Job is running;
-	// - "Completed": The Job finished running.
-	Phase JobExecutionPhase `json:"phase"`
+	// Represents the observations of a JobExecution's state. The state of the JobExecution is tied to the Job it manages.
+	// Conditions.type are: "Waiting", "Running", "Succeeded".
+	// Conditions.status are one of True, False, Unknown.
+	// Conditions.reason defines a camelCase expected values and meanings for this field.
+	// Conditions.Message is a human readable message indicating details about the transition.
+
+	// Conditions store the status conditions of a JobExecution.
+	// +operator-sdk:csv:customresourcedefinitions:type=status
+	Conditions []metav1.Condition `json:"conditions,omitempty" patchStrategy:"merge" patchMergeKey:"type" protobuf:"bytes,1,rep,name=conditions"`
 
 	// Job has a reference to the Job from this execution.
 	// +optional
 	Job corev1.ObjectReference `json:"job,omitempty"`
 }
 
-// JobExecutionPhase describes the current state of the Job.
-type JobExecutionPhase string
+// JobExecutionConditionType describes the observed state of a JobExecution and its Job.
+type JobExecutionConditionType string
 
 const (
-	// JobTemplate does not exist.
-	JobExecutionInvalidPhase JobExecutionPhase = "Invalid"
-	// Job is waiting to be scheduled.
-	JobExecutionWaitingPhase JobExecutionPhase = "Waiting"
-	// Job is running.
-	JobExecutionActivePhase JobExecutionPhase = "Active"
-	// Job has finished running.
-	JobExecutionCompletedPhase JobExecutionPhase = "Completed"
-	// Job has failed executing.
-	JobExecutionFailedPhase JobExecutionPhase = "Failed"
+	JobExecutionWaiting   JobExecutionConditionType = "Waiting"
+	JobExecutionRunning   JobExecutionConditionType = "Running"
+	JobExecutionSucceeded JobExecutionConditionType = "Succeeded"
 )
 
 //+kubebuilder:object:root=true